ai/control/agent-ops/model-evaluation/gen-ai-evaluations/assets/evaluation-scripts/requirements.txt
ibm-watsonx-gov[metrics]>=1.4.2,<2.0.0
pandas>=2.1.0
python-dotenv>=1.0.0
